requirements.txt
fastapi>=0.75.0
uvicorn[standard]>=0.17.0
python-dotenv>=0.20.0
google-adk>=0.1.0
litellm
google-generativeai # Explicitly add if needed for types, ADK usually includes it